# server/openapi/v1.yaml
openapi: 3.0.0
info:
  title: Immigration Horizons Staff API
  version: "1.0"
  description: Canonical JSON API for the Angular Enterprise UI.
servers:
  - url: /api/v1
paths:
...
  /staff/cases:
    get:
      summary: List cases with search and stage filters
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: stage
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated cases
  /staff/cases/{id}:
    get:
      summary: Get case details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Case details
        '404':
          description: Case not found or inaccessible
...
  /staff/cases/{caseId}/members:
    post:
      summary: Add employee team member to case workspace
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId]
              properties:
                employeeId: { type: string }
                role: { type: string, default: contributor }
      responses:
        '200':
          description: Updated case details
...
  /staff/cases/{caseId}/member-options:
    get:
      summary: Get active employees selectable for team assignment
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of active employees
...
  /staff/cases/{id}/tasks:
    get:
      summary: List tasks for a case
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of tasks
    post:
      summary: Create a case-native task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                type: { type: string }
                priority: { type: string }
                dueDate: { type: string, format: date-time, nullable: true }
                assignee: { type: string, nullable: true }
                description: { type: string }
      responses:
        '200':
          description: Success
...